+++ src/pubsub/channel_subscriptions.rs
@@ -21,7 +21,8 @@ impl_de_ser!(
 
 impl pubsub::Topic for ChannelSubscribeEventsV1 {
     #[cfg(feature = "twitch_oauth2")]
-    const SCOPE: &'static [twitch_oauth2::Scope] = &[twitch_oauth2::Scope::ChannelSubscriptions];
+    const SCOPE: &'static [twitch_oauth2::Scope] =
+        &[twitch_oauth2::Scope::ChannelReadSubscriptions];
 
     fn into_topic(self) -> pubsub::Topics { super::Topics::ChannelSubscribeEventsV1(self) }
 }