+++ CHANGELOG.md
@@ -51,6 +51,8 @@
 - Added `Delete Channel Stream Schedule Segment` that was accidentally not included in `0.6.0`
 - Added `sort` and `id` to `Get Custom Reward Redemption`
 - Added contribution type `OTHER` for Get Hype Train Events
+- Added `Send a Shoutout` endpoint
+- Added `channel.shoutout.receive` and `channel.shoutout.create` eventsub events
 
 ### Fixed